fix: resolve CI failures on #2124 (attempt 1)
